Add The Big Short and The Great Mortality books
When it comes to preserving information and potentially rebuilding civilization, our lows can be equally, if not more, important than our highs. I added these two books as they detail huge, negative events to a large swath of human civilization, how we recovered from them, and how they could be avoided or minimized in the future.